requirements.txt
# ── Core LLM & RAG ─────────────────────────────────────────────────────────
langchain==0.2.16
langchain-community==0.2.16
langchain-openai==0.1.23
openai==1.40.0

# ── Embeddings & Reranking ──────────────────────────────────────────────────
sentence-transformers==3.0.1
transformers==4.44.2
torch==2.4.0

# ── Vector Database ─────────────────────────────────────────────────────────
qdrant-client==1.11.0

# ── Document Parsing ────────────────────────────────────────────────────────
unstructured[pdf,docx]==0.15.0
pdfminer.six==20231228
python-docx==1.1.2
markdown-it-py==3.0.0
beautifulsoup4==4.12.3

# ── API & Frontend ──────────────────────────────────────────────────────────
fastapi==0.112.1
uvicorn[standard]==0.30.6
streamlit==1.37.1
pydantic==2.8.2
pydantic-settings==2.4.0

# ── Async ───────────────────────────────────────────────────────────────────
httpx==0.27.0

# ── Utilities ───────────────────────────────────────────────────────────────
python-dotenv==1.0.1
loguru==0.7.2
tenacity==8.5.0
tqdm==4.66.5
numpy==1.26.4
pandas==2.2.2

# ── Evaluation ──────────────────────────────────────────────────────────────
ragas==0.1.21
datasets==2.21.0

# ── Testing ─────────────────────────────────────────────────────────────────
pytest==8.3.2
pytest-asyncio==0.23.8
